MiszMasz
Security - admin hasło admin/ user hasło user
w car service do metody save dodana notka "Samochód został dodany pomyślnie"
RentalService validacja daty - zwrot zawsze będzie "DZISIAJ" czyli w dzień zwrotu czyli data musi się równać LocalDate.now (chyba XD )
ReservationService dodana validacja na wypadek gdyby auto było już zarezerwowane w dni które wybierzemy. Wyskoczy info że daty sie nie zgadzają
ReturnService - to samo co RentalService z datami
CarForm - dodane opisy "To pole jest wymagane"
CustomerForm dodane pozostałe pola do tworzenia customera
DepartmentForm - pola oznaczone jako wymagane do stworzenia oddziału
RegisterView - wszystkie pola wymagane do rejestracji. Po rejestracji przekierowanie na liste samochodów
RentalForm - oznaczenie wymaganych pól
RezervationForm - oznaczenie wymaganych pól/ info o rezerwacji zakończonej sukcesem/ info o rezerwacji nieudanej
Dodane @Secured("ROLE_ADMIN) do niektórych klas ale to nwm czy aż tak dobrze działa ;) pozdrówki

+++ src/main/java/com/example/rentalcars/service/RentalService.java
@@ -3,9 +3,11 @@ package com.example.rentalcars.service;
 import com.example.rentalcars.enums.CarStatus;
 import com.example.rentalcars.model.RentalModel;
 import com.example.rentalcars.repository.RentalRepository;
+import com.vaadin.flow.component.notification.Notification;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 
 @Service
@@ -17,6 +19,11 @@ public class RentalService {
     private final CarService carService;
 
     public void postAddRental(RentalModel rental) {
+        if(!rental.getDateOfRental().isEqual(LocalDate.now())){
+            Notification.show("Coś napiedaczyłeś z datą drogi kolego");
+            return;
+        }
+
         departmentService.removeCarFromDepartment(rental.getReservation().getCar().getId(), rental.getReservation().getReceptionVenue().getId());
         carService.setCarStatus(rental.getReservation().getCar().getId(), CarStatus.HIRED);
         rentalRepository.save(rental);
